# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/discounts/{id}": {
            "get": {
                "tags": [
                    "Discount"
                ],
                "summary": "Display the specified Discount",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Discount"
                        }
                    },
                    "404": {
                        "description": "Discount not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Discount"
                ],
                "summary": "Update the specified Discount in storage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount updated successfully",
                        "schema": {
                            "$ref": "#/definitions/Discount"
                        }
                    },
                    "404": {
                        "description": "Discount not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Discount"
                ],
                "summary": "Remove the specified Discount from storage",
                "description": "Delete Discount",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Discount id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Discount deleted successfully"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Discount": {
            "required": [
                "store_id",
                "date_from",
                "date_to",
                "description",
                "discount"
            ],
            "properties": {
                "id": {
                    "description": "Discount id",
                    "type": "integer",
                    "format": "int32"
                },
                "store_id": {
                    "description": "Store id",
                    "type": "integer",
                    "format": "int32"
                },
                "date_from": {
                    "description": "Discount start date",
                    "type": "string",
                    "format": "date"
                },
                "date_to": {
                    "description": "Discount end date",
                    "type": "string",
                    "format": "date"
                },
                "description": {
                    "description": "Discount description",
                    "type": "string"
                },
                "limit_discount": {
                    "description": "Limit discount",
                    "type": "number",
                    "format": "double"
                },
                "category_id": {
                    "description": "Category id",
                    "type": "integer",
                    "format": "int32"
                },
                "product_id": {
                    "description": "Product id",
                    "type": "integer",
                    "format": "int32"
                },
                "discount": {
                    "description": "Discount amount",
                    "type": "number",
                    "format": "double"
                },
                "active": {
                    "description": "Active flag",
                    "type": "boolean"
                }
            }
        },
...
    }
}